Change inventory models to approximate blocks.
Being able to tell between the climbable and normal versions of blocks
in world is nice, but it's less useful if you can't tell in your
inventory. Changes the bars on the fences to be the iron block instead
of actually splitting up the model, because it's still relatively clear
like that and doesn't actually need 8 more cubes. Adds a custom texture
that tries to look like the 3d grips on the climbable bars, because I
want it to look like the vanilla bars icon. Also includes the xcf for
the texture.